@@ -25,15 +25,15 @@
     "@mantine/notifications": "^7.13.3",
     "@storybook/addon-actions": "^8.3.5",
     "@storybook/test": "^8.3.5",
-    "@tanstack/react-query": "^5.59.16",
+    "@tanstack/react-query": "^5.60.5",
     "@tanstack/react-query-next-experimental": "^5.59.19",
     "@vercel/speed-insights": "^1.1.0",
     "dayjs": "^1.11.13",
     "embla-carousel-react": "^7.1.0",
     "framer-motion": "^11.11.9",
-    "next": "14.2.15",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^14.2.15",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
     "react-hook-form-persist": "^3.0.0",
     "react-intersection-observer": "^9.13.1",
