@@ -1,10 +1,12 @@
 import Image from 'next/image';
 import Link from 'next/link';
-import { useAuthStore } from '@/src/store/use-auth-store';
+import { usePathname } from 'next/navigation';
+import { useAuth } from '@/src/hooks/use-auth';
 import ImgHeroCrew from '@/public/assets/icons/ic-dumbbell.svg';
 
 export default function HeroCrew() {
-  const { isAuth } = useAuthStore();
+  const currentPath = usePathname();
+  const { isAuth } = useAuth();
 
   return (
     <section className="cursor-default pb-3 md:pb-4 lg:pb-12.5">
@@ -21,7 +23,7 @@ export default function HeroCrew() {
           </p>
         </div>
         <Link
-          href={isAuth ? '/crew/create' : '/login'}
+          href={isAuth ? '/crew/create' : `/login?redirect=${currentPath}`}
           className="btn-filled -translate-y-1 items-center rounded-xl px-4 py-2 text-sm font-semibold md:h-11 md:text-lg md:font-bold"
         >
           크루 만들기
